Disabling the mal-formatted GeoIP db example
It seems the GeoIP gem will not let me know
if a db file is a proper one.
So I can't really test this, and it's not in
GeoRedirect's scope anyway.
In any case, we will catch such improper db
file later in the run, when calling
`GeoIP::country` will raise (and we already
catch this error).